Handle changing display modes better
