+++ MooToo/ui/constants.py
@@ -1,0 +1,12 @@
+from enum import Enum, auto
+
+
+#####################################################################################################
+class DisplayMode(Enum):
+    GALAXY = auto()
+    PLANET = auto()
+    ORBIT = auto()
+    SCIENCE = auto()
+    FLEET = auto()
+    COLONY_SUM = auto()
+    PLANET_SUM = auto()
